Make build.js be in-charge of generation
