+++ packages/nimble-tokens/.eslintignore
@@ -1,3 +1,2 @@
 dist/*
 dist/styledictionary/*
-source/styledictionary/build.js
