+++ packages/nimble-tokens/source/styledictionary/config.js
@@ -1,46 +1,5 @@
 const StyleDictionary = require('style-dictionary');
 
-module.exports = {
-    source: [
-        'properties/colors.json',
-        'properties/fonts.json',
-        'properties/custom.json',
-        'properties/sizes.json'
-    ],
-    platforms: {
-        css: {
-            files: [
-                {
-                    destination: 'variables.css',
-                    format: 'css/variables'
-                }
-            ],
-            transformGroup: 'css',
-            buildPath: '../../dist/styledictionary/css/'
-        },
-        js: {
-            files: [
-                {
-                    destination: 'tokens.js',
-                    format: 'javascript/es6'
-                }
-            ],
-            transformGroup: 'js',
-            buildPath: '../../dist/styledictionary/js/'
-        },
-        scss: {
-            files: [
-                {
-                    destination: 'variables.scss',
-                    format: 'scss/variables'
-                }
-            ],
-            transformGroup: 'scss',
-            buildPath: '../../dist/styledictionary/scss/'
-        }
-    }
-};
-
 // Lodash code inlined so we don't have to import another library
 
 // getTag.js
